fix: add amount for composition dialog

+++ services/admin-api/src/migrations/1658980520100-create-ownership.ts
@@ -22,7 +22,7 @@ export class CreateOwnershipAt1658980520100 implements MigrationInterface {
         },
         {
           name: "amount",
-          type: "int",
+          type: "uint256",
         },
         {
           name: "created_at",
